hasher: Implement NoHash.
NoHash type fits as a type parameter with Xor8 filter, or any
other filter that is parametrized over a hash-type.

NoHash, as the name suggest, shall panic for filter-api that needs
hashing. Applications are expected to compute the digest outside
the filter, and use only those api that takes the key-digest
as arguments.