fuse8/xor8: use BTreeMap to deduplicate keys.
APIs insert(), populate() and build() accept any key type that
are hashable, this means it is upto the Xor8 and Fuse8 types
to handle duplicate digests. We are using the BTreeMap as the
backing data-structure to dedup incoming digests.

Note that, for build_keys() it is upto the caller to ensure there
are no duplicate keys. In other words build_keys() is the fast
path, bypassing BTreeMap operations, for situations where caller
can find efficient ways to dedup key digests.